Asgaard now conpiles, most of vioarr compiles on linux. Still need to handle key conversion.

+++ libs/asgaard/memory_pool_unix.cpp
@@ -77,7 +77,7 @@ namespace Asgaard {
             return nullptr;
         }
         
-        if (memoryOffset >= m_size) {
+        if ((size_t)memoryOffset >= m_size) {
             return nullptr;
         }
